Split defunc_fold.rs at a clean impl-block boundary to clear max-lines (1016 -> 604+413)
